refactor: refactor orderController error handling and data types
- Change translation for &#34;OrderNotFound&#34; in en.yaml and fa.yaml
- Change data type of &#34;Code&#34; field in BaseApiResponse from int64 to uint
- Change entity type in orderController.all function from entities.OrderEntity to entities.Order
- Change error message and error code in orderController.all function&#39;s error handling

+++ presentation/http/api/v1/controllers/base.go
@@ -32,11 +32,11 @@ type BaseApiResponse[TD any] struct {
 	err      error
 	httpCode int
 	Message  string `json:"message"`
-	Code     int64  `json:"code" format:"int64"`
+	Code     uint   `json:"code" format:"int64"`
 	Data     TD     `json:"data"`
 } //@name BaseApiResponse
 
-func newBaseApiResponse[TD any](error error, message string, code int64, data TD, httpCode int) BaseApiResponse[TD] {
+func newBaseApiResponse[TD any](error error, message string, code uint, data TD, httpCode int) BaseApiResponse[TD] {
 	return BaseApiResponse[TD]{
 		err:      error,
 		Message:  message,
